refactor: Split config.go handlers into domain-specific files
Move the 74 handler functions out of config.go into 10 new files
grouped by domain, reducing config.go from ~6200 to ~1330 lines.

- config_audio.go: ADEVICE, ARATE, ACHANNELS
- config_channel.go: CHANNEL, MODEM, PTT, DCD, FX25, IL2P, etc.
- config_digi.go: DIGIPEAT, REGEN, FILTER
- config_tt.go: TTPOINT, TTGRID, TTMACRO, etc.
- config_igate.go: IGSERVER, IGLOGIN, IGTXVIA, SATGATE
- config_network.go: AGWPORT, KISSPORT, NULLMODEM, DNSSD
- config_gps.go: GPSNMEA, GPSD, WAYPOINT
- config_beacon.go: BEACON, XBEACON, SMARTBEACON, beacon_options
- config_ax25.go: FRACK, RETRY, PACLEN, MAXFRAME, NOXID
- config_misc.go: LOGDIR, LOGFILE
